Demo5 b26a release
- Added a feature to incorporate receiver estimates of code and phase
stdevs into measurement variances (pos2-rcvstds=on/off)
-Modified encoding of stdevs into the RINEX observation files to a
standard format for all receivers
- Added a feature to exclude a single satellite from ambiguity
resolution every epoch for which the number of satellites is larger than
a configurable limit (pos2-mindropsats).  Each epoch a different
satellite is excluded.  If excluding the satellite results in a
significant improvement in the AR ratio, then that satellite is removed
from the list of satellites used for AR.
- Modified static-start solution mode to return to static and reset the
position if the rover has been inactive for more than 5 minutes
- Minor clean-up