intpres: return the minimum absolute time delta
The returned time delta, between the rover time and the base
observation time, is used as an input to varerr() to contribute to
estimating the measurement variance. When the rover time is close to a
base observation time the residual interpolation error is expected to
be lower and the minimum absolute time delta is also lower.