estpos: correct array allocations with QZSDT enabled
Fix memory corruption with QZSDT enabled and only one system configured.

rescode() can include up to (NX-3) extra constraints, and estpos() needs to
allocate memory for all of these in the worst case, and was allocating an
extra 4 rows but needs 5 extra rows with QZSDT enabled. Allocate an extra
(NX-3) so that this handles more constraints being added in future, handles an
increase in NX.