stream writeserial: patch writes for non-WIN32 platforms
The non-WIN32 path was always returning 0 indicating an error.

Further it was flagging an error if not all the buffer was written.

The serial connection is opened non-blocking. So the write can return
less than the full amount requested. This function appears intended to
be non-blocking, so in this case return the amount written rather than
flagging an error.

This is not a complete fix which might be more extensive. The write
can also return zero which is not necessarily an error, and it might
also return -1 and set errno to EAGAIN if it would have blocked.