time_str: remove this non-reentrant function, replace with time2str
time2str() now returns a pointer to the string buffer which makes
it an easier replacement for the non-reentrant function time_str().

In particular time_str() was often used in trace function arguments
and if is useful to be able to avoid generating the time with not
emitting the trace output to keep these as trace function arguments.