postpos inputobs: guard against iobsr being out of range
The iobsr index can be out of range of the obs data, beyond the end or
the start, such as when there is only one receiver. inputobs() was
indexing into the data using the out of range iobsr index.

Remove a redundant call to nextobsf() in the case that an immediate
prior call returned zero.

Small performance improvement for nextobsf() and nextobsb(). Avoid
comparing the obs times if an obs with a different rcv is found first.