Qt gui monitor: nav empty guard, widths, rtcm ssr
Rework the guarding of the nav eph. The indexing here needs to
multiply the offset by the number of sats, see rtksvr.c
update_eph(). Also use the nav eph number of element to guard the
offset index and fill with a zero eph if oob. The demo5 code has moved
to index per sat and then by the offset, and has a per sat number of
elements, but the code is similiar.

The width elements were short in setNavigation() and setSat() causing
OOB access. Just add the extra entries for now but the cause might
need review.

In showRtcmSsr(), slightly rework an iteration over the code indexes
to avoid CODE_NONE, although it was not broken and code2obs() does
catch an OOB code index.