strsvr_qt: set the tcp option history before setting the path
Otherwise setting the history clears the current path address and if
there is no history then the address is cleared. Simply opening the
app and exiting will clear this history, which might be another
matter.

Also ntripClientOptions() was calling setHistory() MAXHIST times but
this function is intended to be called just once.