stream readntrip: correct reading from the response buffer
For the ntrip stream, initial content is read into the response buffer
for processing, but once processed there can be some content remaining
that can be general input from this stream. The code was using the
most recent number of requested bytes and discarding the
reset. Typically the caller was reading into a large buffer so
requesting more than buffered and so all the content was
consumed. Update the code to also consume all the buffer content even
if the caller is using a small buffer. Also expand the two cases to
make the code clearer.

Rework reqntrip_c() and rspntrip_s() a little, to use memmove() rather
than a for() loop, and comment the buffer flushing, just for clarity.