misc udp stream fixes
stream.c: fix UDP trace format, was passing an int for a string
pointer.

rtkrcv: correct the stream types array of strings; can add 'UDP
client' as an output option.

str2str: add 'UDP client' as a stream output option.

QT str2str: correct the size of the paths[] array, there are 7 paths
not 4, correcting an OOB access. Save and load all 7 path options to
the ini file. Correct an outputTypes[] array which was missing an
entry for UDP client causing an OOB for STR_FILE. Correct an index
into the paths for the input options, the UDP input option is index 4.