Fix of inconsistent data from Septentrio receivers
Move the check to flush the collected measurement data from the previous
epoch to the beginning of the reading routines to avoid inconsistent
data sets.

Before, the first data from a new measurement epoch would already be
stored in the output buffer together with the data of the previous epoch
before the whole data was flushed.

The current ode requires a global variable to track the current
measurement epoch which is, unfortunately, not ideal.