eph.fit: improve handling, QZSS
Add a sys argument to decode_frame so that it can specialise for QZSS
where the interpretation of the fit bit differs from GPS.

Rework some usage of the eph.fit which is defined in RTKLib to be in
hours, so the decoded fit time, not the binary bit. It was not being
consistently set. Now it is set to 4hrs for GPS when appropriate, and
for now to 6hr when over 4hrs (it could be better decoded). For QZSS
it is set to 2hrs when appropriate and to 4hrs when over 2hrs; QZSS
appears to only use 2hr.

The fit time is not currently used by RTKLib, just past through.