rtkpos: detect code changes per frequency index
The code is currently limited to one bias per satellite and pair of
receiver signals. If one of these signals changes then flag a slip to
have the bias reinitialized. Otherwise such changes can throw off the
solution.

The expected observation code is initialized when first seen, and
typically does not change, but the processing is now robust to signal
changes. This initialization accepts different signals at the same
frequency index across a system, but just one signal at a time for a
particular satellite and receiver per frequency index.