rtkpos: rework the rolling exclusion of satellites
The heuristics in manage_amb_LAMBDA() select a rolling satellite to
test for exclusion, to see if it improves the ambiguity ratio. This
was intending to note the sats used for AR, but was also iterating
over the frequencies and noting sat/freq used for AR, but the array to
hold the sat mapping was only large enough for the sats, not
sats*freqs, leading to OOB access when there were many observations
and frequencies. It was rolling over this combination, not as
intended. Further it was sensitive to changes in the set of sats used
for AR, the rolling point could jump around, jump back, and not
progress smoothly thought the set of sats. Rework to address these
issues, to roll through that sats smoothly.