Re-add CMake support
This adds only the CMakeLists.txt files to the repository again. It does
not touch any other files, except of .gitignore.

The update of the #if in rtklib.h to support Visual C will came in a
dedicated commit.

Building was tested with GCC, MinGW and clang. Embarcadero C++ Builder
is also not yet tested/supported (see issue #562).